Split engine unit tests into individual files.
Each file covers a particular feature. These changes are intended to
make the tests easier to navigate.
